some cleanup and add varying penalty strategy
